feat: add overview tab for game stats and roles analysis
- Implemented OverviewTab component to display recent match details, player statistics, and recent matches.
- Removed PlayerDetails component as its functionality is now integrated into OverviewTab.
- Added RolesTab component for role-based performance analysis, including charts and player statistics.
- Enhanced game router to calculate and return role combinations and their statistics.
- Updated win rate calculations to handle cases with zero matches.
